requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-allauth==0.51.0
idna==3.3
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.11

pytest~=7.1.2
py~=1.11.0
colorama~=0.4.4
attrs~=22.1.0
pluggy~=1.0.0
iniconfig~=1.1.1
pexpect~=4.8.0
helpers~=0.2.0
utils~=1.0.1
keyring~=23.5.0
selenium~=4.4.3
djangorestframework~=3.14.0